constpat nearly works again.
--HG--
rename : examples/constpat/Makefile => examples/constpat/Makefile-moved

+++ examples/foobarbaz/bar.c
@@ -1,6 +1,6 @@
-extern int baz(int);
+extern long int baz(long int);
 
-int bar(int a)
+long int bar(long int a)
 {
 	return baz(a-1);
 }
