+++ examples/foobarbaz/baz.c
@@ -1,7 +1,7 @@
-extern int foo(int);
-int bum = 42;
+extern long int foo(long int);
+long int bum = 42;
 
-int baz(int a)
+long int baz(long int a)
 {
 	return foo(a-1);
 }
